Skips games on save that don't have metadata
This will fix long save and exit times for people with big libraries.
The problem comes not from saving all metadata but from games that dont
have metadata.

If a game has no metadata then it also won't be in the gamelist.xml file
so when ES goes to save that game it first tries to find it in the xml
tree by looping through the whole gamelist tree.  The equation of how
many unneeded loops are happening is
loops = (gamecount - gameNotInGamelist) * gameNotInGamelist

So if a system has 705 roms and 31 aren't scraped then:
loops = (705 - 31) * 31
loops = 20,894

So for every game that isn't scraped or in gamelist.xml, it will add to
the amount of cycles that are wasted looking for it in a file it wont be
in anyway.